requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
langchain==0.3.7
langgraph==0.2.28
gradio-client==1.3.0
huggingface-hub==0.24.7
Pillow==10.4.0
python-dotenv==1.0.1
configparser==7.0.0
requests==2.32.3
aiofiles==23.2.1
jinja2==3.1.4      
gradio       
google-generativeai
openai
